+++ src/app/pages/loader-preview/loader-preview.page.ts
@@ -19,6 +19,8 @@ export class LoaderPreviewPage implements OnInit {
   public heightPreviewArea = '183';
   private selectedFiles: any;
   srcData: SafeResourceUrl;
+  public file_extension: string;
+  public file_name: string;
 
   constructor(
     public viewCtrl: ModalController,
@@ -47,7 +49,7 @@ export class LoaderPreviewPage implements OnInit {
     // ---------------------------------------------------------------------
     if (file.type.startsWith("image") && (!file.type.includes('svg'))) {
 
-      console.log('LoaderPreviewPage readAsDataURL file TYPE', file.type);
+      console.log('LoaderPreviewPage USE CASE IMAGE readAsDataURL file TYPE', file.type);
       const reader = new FileReader();
       reader.onloadend = (evt) => {
         const img = reader.result.toString();
@@ -65,8 +67,8 @@ export class LoaderPreviewPage implements OnInit {
     } else if (file.type.startsWith("image") && (file.type.includes('svg'))) {
       // this.previewFiles(file)
 
-      console.log('LoaderPreviewPage readAsDataURL file TYPE', file.type);
-      console.log('LoaderPreviewPage readAsDataURL file ', file);
+      console.log('FIREBASE-UPLOAD LoaderPreviewPage readAsDataURL file TYPE', file.type);
+      console.log('FIREBASE-UPLOAD LoaderPreviewPage readAsDataURL file ', file);
       const preview = document.querySelector('#img-preview') as HTMLImageElement;
 
 
@@ -76,7 +78,7 @@ export class LoaderPreviewPage implements OnInit {
         // convert image file to base64 string
         // const img = reader.result as string;
         const img = reader.result.toString();
-        console.log('LoaderPreviewPage readAsDataURL img ', img);
+        console.log('FIREBASE-UPLOAD USE CASE SVG LoaderPreviewPage readAsDataURL img ', img);
 
         // that.fileSelected = that.sanitizer.bypassSecurityTrustResourceUrl(img);
 
@@ -91,13 +93,25 @@ export class LoaderPreviewPage implements OnInit {
         reader.readAsDataURL(file);
       }
 
-
-
       // ---------------------------------------------------------------------
       // USE CASE FILE
       // ---------------------------------------------------------------------
-    } else if (file.type.startsWith("application")) {
-      console.log('LoaderPreviewPage readAsDataURL file TYPE', file.type);
+      // } else if (file.type.startsWith("application") || file.type.startsWith("video") || file.type.startsWith("audio") ) {
+    } else {
+      console.log('FIREBASE-UPLOAD USE CASE FILE LoaderPreviewPage readAsDataURL FILE ', file);
+      console.log('FIREBASE-UPLOAD USE CASE FILE LoaderPreviewPage readAsDataURL file TYPE', file.type);
+      this.file_extension = file.name.substring(file.name.lastIndexOf('.')+1, file.name.length) || file.name;
+      console.log('FIREBASE-UPLOAD USE CASE FILE LoaderPreviewPage readAsDataURL FILE EXTENSION', this.file_extension);
+      this.file_name = file.name
+      console.log('FIREBASE-UPLOAD USE CASE FILE LoaderPreviewPage readAsDataURL FILE NAME', this.file_name);
+      // if (file.type) {
+      //   const file_type_array = file.type.split('/');
+      //   console.log('FIREBASE-UPLOAD USE CASE FILE LoaderPreviewPage readAsDataURL file_type_array', file_type_array);
+      //   this.file_type = file_type_array[1]
+      // } else {
+      //   this.file_type = file.name.substring(file.name.lastIndexOf('.')+1, file.name.length) || file.name;
+
+      // }
 
       this.createFile();
 
@@ -127,42 +141,40 @@ export class LoaderPreviewPage implements OnInit {
 
   // for svg
 
-  previewFiles(file) {
-
-    console.log('LoaderPreviewPage readAsDataURL file TYPE', file.type);
-    console.log('LoaderPreviewPage readAsDataURL file ', file);
-    const preview = document.querySelector('#img-preview');
-    console.log('LoaderPreviewPage readAsDataURL preview ', preview);
-
-    this.readAndPreview(file, preview)
-  }
+  // previewFiles(file) {
+  //   console.log('LoaderPreviewPage readAsDataURL file TYPE', file.type);
+  //   console.log('LoaderPreviewPage readAsDataURL file ', file);
+  //   const preview = document.querySelector('#img-preview');
+  //   console.log('LoaderPreviewPage readAsDataURL preview ', preview);
+  //   this.readAndPreview(file, preview)
+  // }
 
 
-  readAndPreview(file, preview) {
-    console.log('LoaderPreviewPage readAsDataURL preview HERE 1');
-    // Make sure `file.name` matches our extensions criteria
-    if (/\.(jpe?g|png|gif|svg)$/i.test(file.name)) {
-      var reader = new FileReader();
+  // readAndPreview(file, preview) {
+  //   console.log('LoaderPreviewPage readAsDataURL preview HERE 1');
+  //   // Make sure `file.name` matches our extensions criteria
+  //   if (/\.(jpe?g|png|gif|svg)$/i.test(file.name)) {
+  //     var reader = new FileReader();
 
-      reader.addEventListener("load", function () {
-        var image = new Image();
-        image.height = 100;
-        image.title = file.name;
-        image.src = this.result.toString();
-        console.log('LoaderPreviewPage readAsDataURL preview image.src ', image.src);
-        // preview.appendChild(image);
+  //     reader.addEventListener("load", function () {
+  //       var image = new Image();
+  //       image.height = 100;
+  //       image.title = file.name;
+  //       image.src = this.result.toString();
+  //       console.log('LoaderPreviewPage readAsDataURL preview image.src ', image.src);
+  //       // preview.appendChild(image);
 
-        preview.src(image);
-      }, false);
+  //       preview.src(image);
+  //     }, false);
 
-      reader.readAsDataURL(file);
-    }
+  //     reader.readAsDataURL(file);
+  //   }
 
-    if (file) {
-      [].forEach.call(file, this.readAndPreview);
-    }
+  //   if (file) {
+  //     [].forEach.call(file, this.readAndPreview);
+  //   }
 
-  }
+  // }
 
 
   // NOT USED