+++ src/chat21-core/providers/abstract/presence.service.ts
@@ -13,7 +13,17 @@ export abstract class PresenceService {
   abstract BSLastOnline: BehaviorSubject<any> = new BehaviorSubject<any>(null);
 
   // params
-  abstract tenant = environment.tenant;
+  // abstract tenant = environment.tenant;
+
+  private _tenant: string;
+  public setTenant(tenant): void {
+    this._tenant = tenant;
+  }
+  public getTenant(): string {
+    if (this._tenant) {
+      return this._tenant;
+    } 
+  }
 
   // functions
   abstract initialize(): void;