+++ src/chat21-core/providers/abstract/typing.service.ts
@@ -13,7 +13,16 @@ export abstract class TypingService {
   BSSetTyping: BehaviorSubject<any> = new BehaviorSubject<any>(null);
 
   // params
-  abstract tenant = environment.tenant;
+  private _tenant: string;
+  public setTenant(tenant): void {
+    this._tenant = tenant;
+  }
+  public getTenant(): string {
+    if (this._tenant) {
+      return this._tenant;
+    } 
+  }
+
 
   // functions
   abstract initialize(): void;